Fixed : 재결의견작성 상세 - 중복제거 완료 심의안건정보 상세 - 진행중

+++ src/main/java/egovframework/kapa/file/domain/OpinionFileVO.java
@@ -13,7 +13,7 @@ public class OpinionFileVO {
 
 	  private long seqNo;
 	    private Long decisonId;
-	    private Integer reptSeq;
+	    private Long reptSeq;
 	    private Integer opinionType;
 	    private Integer fileType;
 	    private Integer fileSeq;
@@ -22,7 +22,7 @@ public class OpinionFileVO {
 	    private LocalDateTime regdate;
 	    private LocalDateTime uptdate;
 	    private String fileDescription;
-	    private Integer reptOwnrSeq;
+	    private Long reptOwnrSeq;
 
 	
 }
